New York Fix Ups
Peter's main issuewas that the road felt empty in comparison to everything else, and as such I have added a little bit of detail to the roads. The cars are working and shouldnt do the funky stuff they were doing, and there are pipes around the level that let the player easily move to other areas of the map, significantly reducing the time to move through the city.